requirements.txt
fastapi
uvicorn
python-dotenv
azure-ai-projects==1.0.0b7
azure-identity
